pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "war-rig"
version = "0.1.0"
description = "Multi-agent mainframe documentation system using LangGraph"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "War Rig Team", email = "team@warrig.dev"}
]
keywords = ["mainframe", "documentation", "cobol", "jcl", "langgraph", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]

dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
war-rig = "war_rig.cli:app"

[tool.poetry]
name = "war-rig"
version = "0.1.0"
description = "Multi-agent mainframe documentation system using LangGraph"
authors = ["War Rig Team <team@warrig.dev>"]
readme = "README.md"
packages = [{include = "war_rig"}]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.0"
langchain-anthropic = "^0.2.0"
langchain-core = "^0.3.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
pyyaml = "^6.0"
rich = "^13.0.0"
typer = "^0.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.0.0"
mypy = "^1.8.0"
ruff = "^0.3.0"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
war-rig = "war_rig.cli:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]  # line too long (handled by formatter)

[tool.ruff.lint.isort]
known-first-party = ["war_rig"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
